+++ wms-opc/src/main/java/cn/tpl/opc/commons/dto/result/ExportCushionInfoDTO.java
@@ -37,7 +37,7 @@ public class ExportCushionInfoDTO extends AbsBaseDTO {
      * 开口数
      */
     @ExcelProperty("开口数")
-    private Integer openCount;
+    private Short openCount;
 
     /**
      * 扫码器安装顺序
