+++ wms-opc/src/main/java/cn/tpl/opc/entity/CushionDetailEntity.java
@@ -46,5 +46,5 @@ public class CushionDetailEntity implements Serializable {
     /**
      * 开口数
      */
-    private Integer openCount;
+    private Short openCount;
 }
